+++ CHANGELOG.md
@@ -5,7 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [3.4.0] - Unreleased
 
 ### Added
-- 2019-03-28 : (PLB2B-249)  : Added telematics interface
+- 2019-04-02 : (PLB2B-249)  : Added telematics interface
 ### Changed
 ### Fixed
 ### Removed